+++ Sources/ApplicativeRouter/SyntaxRouter.swift
@@ -1,45 +1,15 @@
-import Either
 import Foundation
 import Prelude
 import Optics
-
-// todo: move to prelude: right associative applicative
-infix operator <%>: infixr4
-infix operator %>: infixr4
-infix operator <%: infixr4
-
-// todo: move to prelude
-public func >-> <A, B, C>(lhs: @escaping (A) -> B?, rhs: @escaping (B) -> C?) -> (A) -> C? {
-  return lhs >>> flatMap(rhs)
-}
-
-private struct RequestData: Monoid {
-  var method: Method? = .get
-  var path: [String] = []
-  var query: [String: String] = [:]
-  var body: Data? = nil
-
-  static var empty = RequestData()
-
-  static func <>(lhs: RequestData, rhs: RequestData) -> RequestData {
-    return .init(
-      method: lhs.method ?? rhs.method,
-      path: lhs.path + rhs.path,
-      query: lhs.query.merging(rhs.query, uniquingKeysWith: { $1 }),
-      // todo: is coalescing enough or should we be appending?
-      body: lhs.body ?? rhs.body
-    )
-  }
-}
 
 // MARK: - Syntax Router
 
 // TODO: should this be generic over any monoid `M` instead of using `RequestData` directly?
 // TODO: generic over an error semigroup too?
 public struct Router<A> {
-  fileprivate let parse: (RequestData) -> (rest: RequestData, match: A)?
-  fileprivate let print: (A) -> RequestData?
-  fileprivate let template: (A) -> RequestData?
+  let parse: (RequestData) -> (rest: RequestData, match: A)?
+  let print: (A) -> RequestData?
+  let template: (A) -> RequestData?
 
   public func match(request: URLRequest) -> A? {
     return (self <% _end).parse(route(from: request))?.match
@@ -152,173 +122,6 @@ extension Router {
   }
 }
 
-// Combinators
-
-/// Processes and consumes a single path component matching the string provided.
-public func lit(_ str: String) -> Router<Prelude.Unit> {
-  return Router<Prelude.Unit>(
-    parse: { route in
-      uncons(route.path).map { _, ps in
-        (.init(method: route.method, path: ps, query: route.query, body: route.body), unit)
-      }
-    },
-    print: { _ in .init(method: nil, path: [str], query: [:], body: nil) },
-    template: { _ in .init(method: nil, path: [str], query: [:], body: nil) }
-  )
-}
-
-/// Processes and consumes a path component and tries to convert it to type `A` using the partial isomorphism
-/// supplied.
-public func pathParam<A>(_ f: PartialIso<String, A>) -> Router<A> {
-  return Router<A>(
-    parse: { route in
-      guard let (p, ps) = uncons(route.path), let v = f.image(p) else { return nil }
-      return (RequestData(method: route.method, path: ps, query: route.query, body: route.body), v)
-    },
-    print: { a in
-      .init(method: nil, path: [f.preimage(a) ?? ""], query: [:], body: nil)
-    },
-    template: { a in
-      .init(method: nil, path: [":\(typeKey(a))"], query: [:], body: nil)
-  })
-}
-
-/// Processes (and does not consume) a query param keyed by `key` as a string.
-public func queryParam(_ key: String) -> Router<String> {
-  return queryParam(key, .id)
-}
-
-/// Processes (and does not consume) a query param keyed by `key`, and then tries to convert it to type `A`
-/// using the partial isomorphism supplied.
-public func queryParam<A>(_ key: String, _ f: PartialIso<String, A>) -> Router<A> {
-  return .init(
-    parse: { route in
-      guard let str = route.query[key] else { return nil }
-      return f.image(str).map { (route, $0) }
-    },
-    print: { a in
-      RequestData(method: nil, path: [], query: [key: f.preimage(a) ?? ""], body: nil)
-    },
-    template: { a in
-      RequestData(method: nil, path: [], query: [key: ":\(typeKey(a))"], body: nil)
-  })
-}
-
-/// Processes the body data of the request.
-public let dataBody = Router<Data>(
-  parse: { route in route.body.map { (route, $0) } },
-  print: { .init(method: nil, path: [], query: [:], body: $0) },
-  template: { .init(method: nil, path: [], query: [:], body: $0) }
-)
-
-/// Processes the body data of the request as a string.
-public let stringBody = dataBody.map(PartialIso.data.inverted)
-
-/// Processes the body data of the request into form data of (key, value) pairs.
-public let formDataBody = stringBody.map(.formEncodedFields)
-
-public func formField(_ name: String) -> Router<String> {
-  return formDataBody.map(key(name))
-}
-
-public func formFields(_ names: String...) -> Router<[String: String]> {
-  return formDataBody.map(keys(names))
-}
-
-public func formDataBody<A: Codable>(_ type: A.Type) -> Router<A> {
-  return ApplicativeRouter.formDataBody
-    .map(jsonDictionaryToData >>> PartialIso.codableToData.inverted)
-}
-
-/// Parses the body data of the request as JSON and then tries to decode the data into a value of type `A`.
-public func jsonBody<A: Codable>(_ type: A.Type) -> Router<A> {
-  return dataBody.map(PartialIso.codableToData.inverted)
-}
-
-/// Parses the end of the request data by making sure that all of the path components have been consumed.
-public let _end = Router<Prelude.Unit>(
-  parse: { route in
-    route.path.isEmpty
-      ? (RequestData(method: route.method, path: [], query: [:], body: nil), unit)
-      : nil
-  },
-  print: const(.empty),
-  template: const(.empty)
-)
-
-extension Router {
-  /// Parses and consumes a path component as an int.
-  public static var int: Router<Int> { return pathParam(.int) }
-
-  /// Parses and consumes a path component as a bool.
-  public static var bool: Router<Bool> { return pathParam(.bool) }
-
-  /// Parses and consumes a path component as a string.
-  public static var string: Router<String> { return pathParam(.id) }
-
-  /// Parses and consumes a path component as a double.
-  public static var num: Router<Double> { return pathParam(.double) }
-}
-
-/// Parses the query params to create a value of type `A` via the `Codable` protocol.
-/// TODO: this only works for types `A` with string fields. Is it possible to improve that?
-public func queryParams<A: Codable>(_ type: A.Type) -> Router<A> {
-  return .init(
-    parse: { route in
-      (try? JSONSerialization.data(withJSONObject: route.query))
-        .flatMap { try? JSONDecoder().decode(A.self, from: $0) }
-        .map { (route, $0) }
-    },
-    print: { a in
-      let params = (try? JSONEncoder().encode(a))
-        .flatMap { try? JSONSerialization.jsonObject(with: $0) }
-        .flatMap { $0 as? [String: Any] }
-        .map { $0.mapValues { "\($0)" } }
-        ?? [:]
-      return RequestData(method: nil, path: [], query: params, body: nil)
-    },
-    template: { a in
-      let params = (try? JSONEncoder().encode(a))
-        .flatMap { try? JSONSerialization.jsonObject(with: $0) }
-        .flatMap { $0 as? [String: Any] }
-        .map { $0.mapValues { _ in ":string" } }
-        ?? [:]
-      return RequestData(method: nil, path: [], query: params, body: nil)
-  })
-}
-
-/// Parses the HTTP method verb of the request.
-public func method(_ method: Method) -> Router<Prelude.Unit> {
-  return Router(
-    parse: { route in
-      route.method == method
-        ? (route, unit)
-        : nil
-    },
-    print: { _ in  .init(method: method, path: [], query: [:], body: nil) },
-    template: { _ in  .init(method: method, path: [], query: [:], body: nil) }
-  )
-}
-
-public let delete = method(.delete)
-public let get = method(.get)
-public let head = method(.head)
-public let options = method(.options)
-public let patch = method(.patch)
-public let post = method(.post)
-public let put = method(.put)
-
-// MARK: - Experimental
-
-// TODO: open for discussion
-infix operator <=>: infixr9
-
-public func <=> <A> (lhs: String, rhs: PartialIso<String, A>) -> Router<A> {
-  return queryParam(lhs, rhs)
-}
-
-// MARK: - Private
-
 private func route(from request: URLRequest) -> RequestData {
   let method = request.httpMethod.flatMap(Method.init(string:)) ?? .get
 
@@ -348,26 +151,4 @@ private func request(from route: RequestData) -> URLRequest? {
   request?.httpMethod = route.method?.rawValue
   request?.httpBody = route.body
   return request
-}
-
-private func typeKey<A>(_ a: A) -> String {
-  // todo: convert camel case to snake case?
-  let typeString = "\(type(of: a))"
-  let typeKey: String
-  if typeString.contains("Optional<") {
-    typeKey = "optional_\(typeString)"
-      .replacingOccurrences(of: "Optional<", with: "")
-      .replacingOccurrences(of: ">", with: "")
-      .lowercased()
-  } else if typeString.contains("Either<") {
-    typeKey = "\(typeString)"
-      .replacingOccurrences(of: "Either<", with: "")
-      .replacingOccurrences(of: ", ", with: "_or_")
-      .replacingOccurrences(of: ">", with: "")
-      .lowercased()
-  } else {
-    typeKey = typeString.lowercased()
-  }
-
-  return typeKey
 }
