+++ Sources/ApplicativeRouter/Todo.swift
@@ -1,0 +1,11 @@
+import Prelude
+
+// todo: move to prelude: right associative applicative
+infix operator <%>: infixr4
+infix operator %>: infixr4
+infix operator <%: infixr4
+
+// todo: move to prelude
+public func >-> <A, B, C>(lhs: @escaping (A) -> B?, rhs: @escaping (B) -> C?) -> (A) -> C? {
+  return lhs >>> flatMap(rhs)
+}
